+++ src/tensors/abstracttensor.jl
@@ -38,12 +38,30 @@ function InnerProductStyle(::Type{TT}) where {TT <: AbstractTensorMap}
     return InnerProductStyle(spacetype(TT))
 end
 
+# storage types and promotion system
+# ----------------------------------
 @doc """
     storagetype(t::AbstractTensorMap) -> Type{A<:AbstractVector}
     storagetype(T::Type{<:AbstractTensorMap}) -> Type{A<:AbstractVector}
 
 Return the type of vector that stores the data of a tensor.
+If this is not overloaded for a given tensor type, the default value of `storagetype(scalartype(t))` is returned.
+
+See also [`similarstoragetype`](@ref).
 """ storagetype
+storagetype(t) = storagetype(typeof(t))
+function storagetype(::Type{T}) where {T <: AbstractTensorMap}
+    if T isa Union
+        # attempt to be slightly more specific by promoting unions
+        Ma = storagetype(T.a)
+        Mb = storagetype(T.b)
+        return promote_storagetype(Ma, Mb)
+    else
+        # fallback definition by using scalartype
+        return similarstoragetype(scalartype(T))
+    end
+end
+storagetype(T::Type) = throw(MethodError(storagetype, T))
 
 # storage type determination and promotion - hooks for specializing
 # the default implementation tries to leverarge inference and `similar`
@@ -69,6 +87,8 @@ appropriate storage types. Additionally this registers the default storage type 
     used in constructor-like calls, and therefore will return the exact same type for a `DenseVector`
     input. The latter is used in `similar`-like calls, and therefore will return the type of calling
     `similar` on the given `DenseVector`, which need not coincide with the original type.
+
+See also [`promote_storagetype`](@ref).
 """ similarstoragetype
 
 # implement in type domain
@@ -101,6 +121,74 @@ similarstoragetype(::Type{D}, ::Type{T}) where {D <: AbstractDict{<:Sector, <:Ab
 
 # default storage type for numbers
 similarstoragetype(::Type{T}) where {T <: Number} = Vector{T}
+
+@doc """
+    promote_storagetype([T], A, B, C...)
+    promote_storagetype([T], TA, TB, TC...)
+
+Determine an appropriate storage type for the combination of tensors `A` and `B`, or tensors of type `TA` and `TB`.
+Optionally, a scalartype `T` for the destination can be supplied that might differ from the inputs.
+""" promote_storagetype
+
+@inline promote_storagetype(A::AbstractTensorMap, B::AbstractTensorMap, Cs::AbstractTensorMap...) =
+    promote_storagetype(storagetype(A), storagetype(B), map(storagetype, Cs)...)
+@inline promote_storagetype(::Type{T}, A::AbstractTensorMap, B::AbstractTensorMap, Cs::AbstractTensorMap...) where {T <: Number} =
+    promote_storagetype(similarstoragetype(A, T), similarstoragetype(B, T), map(Base.Fix2(similarstoragetype, T), Cs)...)
+
+@inline function promote_storagetype(
+        ::Type{A}, ::Type{B}, Cs::Type{<:AbstractTensorMap}...
+    ) where {A <: AbstractTensorMap, B <: AbstractTensorMap}
+    return promote_storagetype(storagetype(A), storagetype(B), map(storagetype, Cs)...)
+end
+@inline function promote_storagetype(
+        ::Type{T}, ::Type{A}, ::Type{B}, Cs::Type{<:AbstractTensorMap}...
+    ) where {T <: Number, A <: AbstractTensorMap, B <: AbstractTensorMap}
+    return promote_storagetype(similarstoragetype(A, T), similarstoragetype(B, T), map(Base.Fix2(similarstoragetype, T), Cs)...)
+end
+
+# promotion system in the same spirit as base/promotion.jl
+promote_storagetype(::Type{Base.Bottom}, ::Type{Base.Bottom}) = Base.Bottom
+promote_storagetype(::Type{T}, ::Type{T}) where {T} = T
+promote_storagetype(::Type{T}, ::Type{Base.Bottom}) where {T} = T
+promote_storagetype(::Type{Base.Bottom}, ::Type{T}) where {T} = T
+
+function promote_storagetype(::Type{T}, ::Type{S}) where {T, S}
+    @inline
+    # Try promote_storage_rule in both orders. Typically only one is defined,
+    # and there is a fallback returning Bottom below, so the common case is
+    #   promote_storagetype(T, S) =>
+    #   promote_storage_result(T, S, result, Bottom) =>
+    #   typejoin(result, Bottom) => result
+    return promote_storage_result(T, S, promote_storage_rule(T, S), promote_storage_rule(S, T))
+end
+
+@inline promote_storagetype(T, S, U) = promote_storagetype(promote_storagetype(T, S), U)
+@inline promote_storagetype(T, S, U, V...) = promote_storagetype(promote_storagetype(T, S), U, V...)
+
+@doc """
+    promote_storage_rule(type1, type2)
+
+Specifies what type should be used by [`promote_storagetype`](@ref) when given values of types `type1` and
+`type2`. This function should not be called directly, but should have definitions added to
+it for new types as appropriate.
+""" promote_storage_rule
+
+promote_storage_rule(::Type, ::Type) = Base.Bottom
+# Define some methods to avoid needing to enumerate unrelated possibilities when presented
+# with Type{<:T}, and return a value in general accordance with the result given by promote_type
+promote_storage_rule(::Type{Base.Bottom}, slurp...) = Base.Bottom
+promote_storage_rule(::Type{Base.Bottom}, ::Type{Base.Bottom}, slurp...) = Base.Bottom # not strictly necessary, since the next method would match unambiguously anyways
+promote_storage_rule(::Type{Base.Bottom}, ::Type{T}, slurp...) where {T} = T
+promote_storage_rule(::Type{T}, ::Type{Base.Bottom}, slurp...) where {T} = T
+
+promote_storage_result(::Type, ::Type, ::Type{T}, ::Type{S}) where {T, S} = (@inline; promote_storagetype(T, S))
+# If no promote_storage_rule is defined, both directions give Bottom => error
+promote_storage_result(T::Type, S::Type, ::Type{Base.Bottom}, ::Type{Base.Bottom}) =
+    throw(ArgumentError("No promotion rule defined for storagetype `$T` and `$S`"))
+
+# promotion rules for common vector types
+promote_storage_rule(::Type{T}, ::Type{S}) where {T <: DenseVector, S <: DenseVector} =
+    T === S ? T : throw(ArgumentError("No promotion rule defined for storagetype `$T` and `$S`"))
 
 # tensor characteristics: space and index information
 #-----------------------------------------------------
@@ -224,8 +312,7 @@ end
 # tensor characteristics: work on instances and pass to type
 #------------------------------------------------------------
 InnerProductStyle(t::AbstractTensorMap) = InnerProductStyle(typeof(t))
-storagetype(t) = storagetype(typeof(t))
-storagetype(T::Type) = throw(MethodError(storagetype, T))
+
 blocktype(t::AbstractTensorMap) = blocktype(typeof(t))
 
 numout(t::AbstractTensorMap) = numout(typeof(t))