+++ src/tensors/tensoroperations.jl
@@ -153,14 +153,8 @@ function TO.tensorcontract_type(
         ::Index2Tuple{N₁, N₂}
     ) where {N₁, N₂}
     S = check_spacetype(A, B)
-    TC′ = promote_permute(TC, sectortype(S))
-    M = promote_storagetype(similarstoragetype(A, TC′), similarstoragetype(B, TC′))
+    M = promote_storagetype(promote_permute(TC, sectortype(S)), A, B)
     return tensormaptype(S, N₁, N₂, M)
-end
-
-# TODO: handle actual promotion rule system
-function promote_storagetype(::Type{M₁}, ::Type{M₂}) where {M₁, M₂}
-    return M₁ === M₂ ? M₁ : throw(ArgumentError("Cannot determine storage type for combining `$M₁` and `$M₂`"))
 end
 
 function TO.tensorcontract_structure(